+++ frontend/src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Routes from "./Routes"
+import Routes from "./routes"
 import { BrowserRouter as Router } from "react-router-dom";
 
 const App = () => {
