+++ frontend/src/Components/index.js
@@ -1,2 +1,4 @@
-export * from "./NewBookingForm";
-export * from "./ClassCard";
+export { default as ClassCard } from "./ClassCard";
+export { default as Footer } from "./Footer";
+export { default as Header } from "./Header";
+export { default as NewBookingForm } from "./NewBookingForm";
