+++ frontend/src/Routes/index.js
@@ -1,7 +1,7 @@
 import React from "react";
-import Home from "../Screen/Home/Home";
-import UpcomingClass from "../Screen/UpcomingClass/UpcomingClass";
-import NewBooking from "../Screen/NewBooking/NewBooking";
+import { Home } from "../screen";
+import { UpcomingClass } from "../screen";
+import NewBooking from "../screen/NewBooking/NewBooking";
 
 import { BrowserRouter as Switch, Route } from "react-router-dom";
