+++ frontend/src/Screen/Home/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Header from "./Header/Header";
-import Footer from "./Footer/Footer";
+import Header from "./Header/Header.jsx";
+import Footer from "./Footer/Footer.jsx";
 import { Link } from "react-router-dom";
 import "./Home.scss";
