+++ frontend/src/Screen/NewBooking/NewBooking.js
@@ -1,9 +1,7 @@
 import React, { useEffect } from "react";
 import { connect } from "react-redux";
-import { Get_Classes } from "../../Redux/Actions";
-import Header from "../../Components/Header/Header";
-import ClassCard from "../../Components/ClassCard";
-import Footer from "../../Components/Footer/Footer";
+import { Get_Classes } from "../../redux/actions";
+import { Header, ClassCard, Footer } from "../../components";
 
 const mapStateToProps = (state) => {
   return { classes: state.ClassReducer.classes };
